finalise split node logic

+++ tests/persistence/toc/read.rs
@@ -12,7 +12,7 @@ fn test_index_table_of_contents_read() {
         .open("tests/data/expected/persist_index_table_of_contents.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(1));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -30,7 +30,7 @@ fn test_index_table_of_contents_read_from_space_index() {
         .open("tests/data/expected/space_index/process_create_node.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -48,7 +48,7 @@ fn test_index_table_of_contents_read_from_space_index_after_insert() {
         .open("tests/data/expected/space_index/process_insert_at.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -66,7 +66,7 @@ fn test_index_table_of_contents_read_from_space_index_with_updated_node_id() {
         .open("tests/data/expected/space_index/process_insert_at_with_node_id_update.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -84,7 +84,7 @@ fn test_index_table_of_contents_read_from_space_index_with_remove_at_node_id() {
         .open("tests/data/expected/space_index/process_remove_at_node_id.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -102,7 +102,7 @@ fn test_index_table_of_contents_read_from_space_index_with_remove_node() {
         .open("tests/data/expected/space_index/process_remove_node.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
@@ -121,7 +121,7 @@ fn test_index_table_of_contents_read_from_space_index_with_create_node_after_rem
         .open("tests/data/expected/space_index/process_create_node_after_remove.wt.idx")
         .unwrap();
     let next_id_gen = Arc::new(AtomicU32::new(2));
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::parse_from_file(
         &mut file,
         0.into(),
         next_id_gen,
