+++ tests/persistence/toc/write.rs
@@ -10,7 +10,7 @@ use crate::{check_if_files_are_same, remove_file_if_exists};
 fn test_persist_index_table_of_contents() {
     remove_file_if_exists("tests/data/persist_index_table_of_contents.wt.idx".to_string());
 
-    let mut toc = IndexTableOfContents::<u8, { INNER_PAGE_SIZE as u32 }>::new(
+    let mut toc = IndexTableOfContents::<u32, { INNER_PAGE_SIZE as u32 }>::new(
         0.into(),
         Arc::new(AtomicU32::new(1)),
     );
